RF | Create BaseFoldersConfigPage, BaseProjectsConfigPage
