+++ src/test/java/school/redrover/FolderTest.java
@@ -276,7 +276,7 @@ public class FolderTest extends BaseTest {
                 .clickCreateAJob()
                 .enterItemName("Mine Project")
                 .selectMultiConfigurationProjectAndOk()
-                .saveConfigurePageAndGoToProjectPage();
+                .clickSaveButton();
 
         Assert.assertTrue(multiPage.getMultiProjectName().getText().contains("Mine Project"));
     }
