+++ src/test/java/school/redrover/NewViewTest.java
@@ -56,7 +56,7 @@ public class NewViewTest extends BaseTest {
                 .setNewViewName(expectedName)
                 .selectListView()
                 .clickCreateButton()
-                .clickViewConfigOkButton()
+                .clickSaveButton()
                 .getViewName();
 
         assertEquals(actualName, expectedName);
@@ -123,7 +123,7 @@ public class NewViewTest extends BaseTest {
                 .selectListView()
                 .clickCreateButton()
                 .selectJobsInJobFilters(folderName1)
-                .clickViewConfigOkButton();
+                .clickSaveButton();
 
         Assert.assertEquals(viewPage.getViewName(), viewName);
         Assert.assertEquals(viewPage.getJobName(folderName1), folderName1);
@@ -156,7 +156,7 @@ public class NewViewTest extends BaseTest {
                 .chooseJobsInJobFilters(folderName1 + " » " + jobName1)
                 .chooseJobsInJobFilters(folderName1 + " » " + jobName3)
                 .chooseJobsInJobFilters(folderName2)
-                .clickViewConfigOkButton();
+                .clickSaveButton();
 
         List<String> actualViewJobsTexts = viewPage.getJobNamesList();
