+++ src/test/java/school/redrover/OrganizationFolderTest.java
@@ -83,7 +83,7 @@ public class OrganizationFolderTest extends BaseTest {
                 .clickNewItem()
                 .enterItemName(organizationFolderName)
                 .selectOrganizationFolderAndOk()
-                .selectDescription("Description")
+                .addDescription("Description")
                 .clickSaveButton()
                 .getTextFromDescription();
