+++ src/test/java/school/redrover/model/OrganizationFolderConfigPage.java
@@ -2,8 +2,9 @@ package school.redrover.model;
 
 import org.openqa.selenium.By;
 import school.redrover.model.base.BaseConfigPage;
+import school.redrover.model.base.BaseFoldersConfigPage;
 
-public class OrganizationFolderConfigPage extends BaseConfigPage<OrganizationFolderConfigPage, OrganizationFolderPage> {
+public class OrganizationFolderConfigPage extends BaseFoldersConfigPage<OrganizationFolderConfigPage, OrganizationFolderPage> {
 
 
     public OrganizationFolderConfigPage(OrganizationFolderPage organizationFolderPage) {
@@ -12,12 +13,6 @@ public class OrganizationFolderConfigPage extends BaseConfigPage<OrganizationFol
 
     public OrganizationFolderConfigPage clickDisable(){
         getDriver().findElement(By.xpath("//label[@data-title='Disabled']")).click();
-
-        return this;
-    }
-
-    public OrganizationFolderConfigPage selectDescription(String description){
-        getDriver().findElement(By.xpath("//textarea[@name='_.description']")).sendKeys(description);
 
         return this;
     }
