+++ src/test/java/school/redrover/model/base/BaseConfigPage.java
@@ -2,6 +2,7 @@ package school.redrover.model.base;
 
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 
 
@@ -27,4 +28,13 @@ public abstract class BaseConfigPage<Self extends BaseConfigPage<?, ?>, JobPage 
         getDriver().findElement(By.xpath("//textarea[contains(@name, 'description')]")).sendKeys(description);
         return (Self) this;
     }
+
+    public WebElement viewDescription(){
+        return getDriver().findElement(By.xpath("//*[@id=\"description\"]/div[1]"));
+    }
+
+    public String getDescription() {
+        return getDriver().findElement(By.xpath("//*[@id='description']/div")).getText();
+    }
+
 }
